Stupid NDK maintainers
For some reason the was split into two parts ndk32 and ndk64. The
stupid part is that they're not responsible for 32 bit respectively 64
bit builds. You need both, gnustl is present in ndk64 builds, however
to be able to build for any target except L you need the ndk32 bit as
well.